Fixed type error + showed labels for conf matrix

+++ musicPlayerWidget.py
@@ -1,9 +1,11 @@
 from scipy.io import wavfile
 from mutagen import mp3
+
 from PyQt5.QtCore import QUrl, pyqtSignal
 from PyQt5.QtGui import QFont
 from PyQt5.QtMultimedia import QMediaPlayer, QMediaContent
 from PyQt5.QtWidgets import QWidget, QLabel, QHBoxLayout, QVBoxLayout
+
 from pyqt_media_slider.mediaSlider import MediaSlider
 from pyqt_svg_button.svgButton import SvgButton
 from PyQt5.QtCore import Qt
@@ -135,7 +137,7 @@ class MusicPlayerWidget(QWidget):
     def __updateDuration(self, duration):
         self.__slider.setRange(0, duration)
         self.__slider.setEnabled(duration > 0)
-        self.__slider.setPageStep(duration / 1000)
+        self.__slider.setPageStep(int(duration / 1000))
         self.durationUpdated.emit(duration)
 
     def setMedia(self, filename, type='mp3'):
